Avg Waiting Time for FCFS, Priority

+++ OpSys/list.h
@@ -12,7 +12,7 @@ public:
   void addPCB(int PID1, int arrivalTime1, int burstTime1, int priority1);
   void addPCB(int pos, int data); //change to data?
   void priorityInsert(int PID1, int arrivalTime1, int burstTime1, int priority1);
-  int calcAvgWaitTime();
+  double calcAvgWaitTime();
   int deletePCB(int PID); //change to data?
   int deletePCB();
   void printVals();
